+++ api/dependencies.py
@@ -1,11 +1,11 @@
-from typing import Annotated, Optional, List
+from typing import Annotated
 from fastapi import Depends, HTTPException
 from sqlalchemy.orm import Session
 from fastapi.security import OAuth2PasswordBearer
-from pydantic import BaseModel
 from decouple import config
 from db.db_connection import get_db
 from auth.auth_handler import decode_token
+from schemas.user import AppUser
 from services.user import UserService
 
 
@@ -17,16 +17,7 @@ OIDC_ROLES_PROPERTY = config("OIDC_ROLES_PROPERTY")
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl=OIDC_TOKEN_ENDPOINT)
 
 
-class AppUser(BaseModel):
-    id: Optional[int] = None
-    username: Optional[str] = None
-    email: Optional[str] = None
-    first_name: Optional[str] = None
-    last_name: Optional[str] = None
-    roles: Optional[List[str]] = None
-
-
-async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)], db: Session = Depends(get_db)):
+async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)], db: Session = Depends(get_db)) -> AppUser:
     decoded = decode_token(token)
     username = decoded[OIDC_USERNAME_PROPERTY]
     user_in_db = UserService(db).get_user(username=username)
