+++ api/schemas/user.py
@@ -1,3 +1,4 @@
+from typing import List, Optional
 from pydantic import ConfigDict, BaseModel
 
 
@@ -16,3 +17,12 @@ class UserGroup(BaseModel):
 class UserRoles(BaseModel):
     role: UserGroup
     model_config = ConfigDict(from_attributes=True)
+
+
+class AppUser(BaseModel):
+    id: Optional[int] = None
+    username: Optional[str] = None
+    email: Optional[str] = None
+    first_name: Optional[str] = None
+    last_name: Optional[str] = None
+    roles: Optional[List[str]] = None
